requirements.txt
aiogram==2.17.1
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.3.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.9
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
pyaes==1.6.1
pyasn1==0.4.8
python-environ==0.4.54
pytz==2021.3
rsa==4.8
Telethon==1.24.0
yarl==1.7.2
